Added tracks recording, displayed tracks list

+++ app/src/main/java/com/darekbx/geotracker/repository/entities/PointDto.kt
@@ -6,8 +6,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "point")
 class PointDto(
-    @PrimaryKey(autoGenerate = true) val id: Int? = null,
-    @ColumnInfo(name = "track_id") val trackId: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long? = null,
+    @ColumnInfo(name = "track_id") val trackId: Long,
     @ColumnInfo(name = "timestamp") val timestamp: Long,
     @ColumnInfo(name = "latitude") val latitude: Double,
     @ColumnInfo(name = "longitude") val longitude: Double,
